[sleeptracker] Added new StatedSleeptracker object
StatedSleeptracker allows to keep track of the sleep status and to
get notified on resumes.

This uses the same idea as mce, where a timerfd on the REALTIME clock
is created and watched. When a resume happens, receiving ECANCELED means
that the realtime clock is getting synced back after suspend - thus
giving us a clue.

Note that despite its name, currently it does NOT track sleep entries,
but only resumes.

Signed-off-by: Eugenio Paolantonio (g7) <me@medesimo.eu>